Rename survey consent file
