+++ packages/surveys/src/consent.spec.ts
@@ -1,4 +1,4 @@
-import { ConsentSurveyPlugin } from "./consent";
+import { ConsentSurveyPlugin } from "./consentSurvey";
 
 test("Does consent survey return chsData correctly?", () => {
   expect(ConsentSurveyPlugin.chsData()).toMatchObject({ chs_type: "consent" });
