+++ packages/surveys/src/index.spec.ts
@@ -1,6 +1,6 @@
 import SurveyPlugin from "@jspsych/plugin-survey";
 import { initJsPsych } from "jspsych";
-import { Trial as ConsentTrial } from "./consent";
+import { Trial as ConsentTrial } from "./consentSurvey";
 import { Trial as ExitTrial } from "./exit";
 import Surveys from "./index";
 import { consentSurveyFunction, exitSurveyFunction } from "./utils";
